add suspense boundary
